requirements_current.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cssselect2==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.69
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langsmith==0.4.6
llvmlite==0.44.0
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
minio==7.2.16
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opencv-python==4.11.0.86
orjson==3.11.0
packaging==25.0
passlib==1.7.4
pillow==11.2.1
playwright==1.54.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.0.3
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.2
rectpack==0.2.2
redis==6.4.0
regex==2024.11.6
reportlab==4.4.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
srt==3.5.3
starlette==0.47.1
structlog==25.4.0
svglib==1.5.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vosk==0.3.45
vulture==2.14
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
